SNAP-150 -- Implemented the global feed settings.
Added support for /rss.xml and /atom.xml on top of the /feed/main/main.(rss|atom).
Added a sub-type to the feed links.
Added support for including a link to the global feed logo image in RSS and Atom files.
Added information about the editor::layout which can be defined in the .conf file (although probably should never be? if not then we should prevent it in the layout.cpp)
Added support for entering a widget geometry. At this time it is limited to the width and height of an image box.
Added support to force the path of an image box widget via the attachment tag. This way we could force the feed default logo to be saved under /feed.
Added global feed settings fields in the dbutils.cpp file.
Allow the dash (-) character in the name of a layout, even though we probably should not.
Prevent the creation of a shortcut on the home page.
Fixed the layout handling of the default layout name. The default found in .conf cannot be a JavaScript but it was viewed that way in one case.
Fixed the owner of the info form to "info" instead of "locale".
Various clean ups.