SNAP-253, SNAP-265: Added a test of the HTTP_HOST in snap.cgi, if it represents an IPv4 or IPv6, err right away.
Enhanced the tracker plugin to also save the resulting HTTP code to a request.
Preven the use of port 0 for the snapserver in snap.cgi.
Fixed the output of errors in snap.cgi so it does not get "chunked out" by Apache.
Added the X-Powered-By header so we know which module in snap generated an error (i.e. snap.cgi or snap_child).
Added a test to verify part of the new IPv4 and IPv6 parsing functions.
Added the "Max-Age" to the cookies so one can use that instead of Expires.
Various clean ups.