SNAP-137: Replaced the use of the q=... query string variable with the use of the REQUEST_URI which works just fine when you try to access the website "properly" (i.e. without /cgi-bin/snap.cgi?...)
Updated the Apache2 configuration example so our users do not end up adding the q=... variable.
Removed the snap_uri() function since it was rendered useless by the removal of the q=... variable.
Removed the definition of the q path in the snapserver.conf since it is not to be used anymore.
Enhanced the error() function in our snap.cgi so the body is different for 404 errors.
Fixed the die() error code for paths that are too long, we wanted a 4XX and there is a 414 for URI too long.
Added the CGI version (apache interface) in our Versions plugin output.