SNAP-253: Added a check in snap.cgi to immediately reply to requests with an empty user agent (it happens from Baidu, at least).
Changed some 503 into 400 so that way we should avoid having more access from that robot (because 5XX means it may work if checked again later.)
Fixed the name of the request method get_name() enum.
Implemented the list of shortcuts.
Fixed the perms() check as for all pages under /admin we always have to use "administer" as the action.