SNAP-168: Started real work on the snapcommunicator.cpp implementation. Added it so it gets compiled.
Fixed the installation of configuration files. We were still missing the new files (snapinit, snapcommunicator) and cleaned up a few things too.
Fixed the syslog name so it does not just say Snap for all logs. That way one knows whether it came from snapinit, snapserver, etc.
Renamed the snapcgi log file by removing the "log" part since the .properties should be more than enough.
Fixed the corresponding debian .install files.
Added support for events from a UDP socket. It has to be different than the TCP connection since it derives from the UDP classes.
Started work on the Snap Communicator message handling (create and parse, to use on both ends of the connection.)
Added a non_blocking() function so we do not have to duplicate that code over and over again.
Moved the parsing of <address>:<ip> info for various servers in a function that everyone can call. The function verifies that the values are at least considered valid.