EX-43: Applied a fix to avoid the &nbsp; entity (and any other lat1 ones... we would need to add the special ones and also all the other languages.)
Added a test to very that the entity replacement works properly.
Make use of the new XSLT parser (helper) class in the layout.cpp file (still need to replace many other instances, though!)
Apply the entity filter when saving HTML in the editor.
Fixed a label for LOG_LEVEL_FATAL, use "fatal error" instead of just "fatal".
Enhanced the QMessageHandler to understand that a word includes a &nbsp; entity so as to allow the XSLT class to fix the problem.
Also made all the message type lowercase so it matches what we have in the log library.