EX-43: Fixed all the instances of QXmlQuery except the one for the sitemap.txt file which is different.
Added a backtrace of 100 items whenever I have a problem with the QXmlQuery() interface, which is likely to help with resolving those problems in the future.
Removed the HTTP_COOKIE print, it was just to debug some potential problems.
Made the backtrace accept an argument so one can specify how far back the stack print out has to go.
Added support for a QDOMDocument in set_xsl(), which later may be useful once we have a QXmlQuery that accepts a DOM as input.
Fixed a bug in the user implementation, it would check != instead of == for the password path and thus return the wrong status.