EX-45: Fixed the user registering and login code to properly support emails with any case.
Added functions to handle the change: for example the get user email now returns the saved email (users::current_email) instead of the user key.
Changed code that handles sessions to make sure we use the user_key and not the original email for those.
Fixed the dual login test so the message "Two Sessions" is given only if you really have another valid session and not just some other session.
Fixed corresponding code in the OAuth2 plugin. Especially, created a function to check whether a user is logged in or not according to his session info.
Added the login redirect variable so the administrator can for a user who logs in to go to another page than his account (assuming there was no other redirection).
Applied a fix to the permissions to retrieve the correct user table row from an email.
Also enhanced the sendmail implementation to get the user_key and not the email directly (all keys need to be canonicalized).
Started work on a password plugin by implementing a user event. It will force users to enter enough of each type of character and not use a password in our blacklist.
Changed implementation of the non_blocking() socket with a call to the event library implementation.
Fixed #include in info.h from form.h to editor.h.
Various clean ups.