SNAP-214: As part of the breaking up User/Form, I now broke up the Sendmail/Editor.
Moved all the unsubscribe support to the Info plugin. The code remains separated in info/unsubscribe.cpp.
Fixed the list output which would prevent the /admin list from showing because it was testing whether the list was view-able and it isn't. We had to test whether it can be administered.
Added the sendmail layout files to the add_library() in CMakeLists.txt.
Fixed the unsubscribe background color as it can be affected by the main theme. Also applied that change to the emails background color.