SNAP-214: Broke up the users plugin in two to avoid circular dependencies.
Moved all the forms (content.xml pages and *-form.xml files) and the code handling them, as well as the default user image to users_ui.
In some cases, the functions were divided, some code was kept in users and other moved to users_ui.
Fixed the login_user() function so the validation_required flag is an out only flag.
Since all the forms are in users_ui, I could also move the sendmail dependency to users_ui (since sendmail depends on users.)
Added a create_logged_in_user_session() function, which is somewhat dangerous, but makes it possible to auto-login a user who just changed their password.
Made the users_ui a default plugin (i.e. it will always be loaded just like users and many others.)
Added the password and users_ui to the documentation and checked, all the loops except "content" and "links" are broken.