SNAP-269: Added support for an editor password field.
Fixed the z-index of the editor toolbar although we probably want to make that dynamic at some point (otherwise it will be under other things like the Feedback button shown on the side.)
Added a sub-list to the editor settings so we can be sent to the test page.
Added a no-paste to fields to prevent Ctrl-V and in some fields. It is not yet implemented in the .js though.
Updated the JavaScript test script so new scripts get checked.
Various clean ups.