SNAP-168: Completely removed the use of libevent from snap_communicator. Not worth the trouble.
Wrote a test to verify some of the implementation of the snap_communicator and snap_communicator_message classes.
Fixed a couple of tests that were inverted in snap_communicator_message.
Renamed "name" as "service" instead.
Use param_name instead of name when referencing the parameter name.
Changed the snap_communicator in a singleton. It simplifies a lot of things and there are really no reason for the need to support more than one event loop.
Moved the read and write capabilities I started writing in the src/snapcommunicator.cpp to the snap_communicator class.
Added a "message" class that interprets the data coming in and going out as snap_communicator_message objects.
Changed the process_signal() with a set of flags to 8 callbacks (plus 2 callbacks in the new buffer and message classes.)
Changed the priority to a simple number. Still limiting that number between 0 and 255 inclusive. This is used to sort the connections out.
Changed the get_events() in a set of functions: is_listener(), is_reader(), is_writer() which works a lot better.
Added support for a timeout on a specific time and date, or as ticks that always happen on a specific scheduled time (no splipping over time).
Added a close() to a couple of classes so one can close the socket without having to immediately destroy the connection.
Properly handles error cases (error, hang up, invalid).
Started the signal implementation, but I did not yet test it.
Added support for enabling and disabling connections on the fly. A disabled connection is not added to the poll() to listen for its events.
Made the necessary changes to the snapwebsites.cpp implementation so it works with all the changes to the snap_communicator environment.
Completely removed the client_impl from snapwebsites.cpp since we were using it only to set KEEPALIVE and close the socket.
Added support for a close() on the bio_client implementation.
Tweaked the code in src/snapcommunicator.cpp so it compiles, but there is otherwise no progress in that one yet.