EX-46: SNAP-67: Added support to parse emails since bounced emails are... emails, in sendmail.
Added a few tests to verify the parsing of emails in sendmail.
Applied a fix to the test plugin implementation which would not set success to false on exceptions.
Added support for == and != with 'char const *' in the QCaseSensitiveString() class.
Renamed the sender and recipient fields because we add an empty line to clearly mark our header and thus we really don't need to have unique names.