SNAP-46: SNAP-67: Changed the column name of bounced emails to use the date first (then the UUID). That way it is sorted by date.
Added support in dbutils so I can verify that the rows are defined as expected.
Fixed the logger setup for TRACE, it now properly uses TRACE instead of INFO as the level when tracing...
Fixed the pattern for the syslog definitions so it more or less matches other servers.
Removed the comments about the fact that certain loggers are not working.
Fixed the configuration that would still use rootLogger instead of logger.snap (otherwise the retrieval fails in our logger implementation.)
Fixed the logger picking at server::instance()->servername() because that may be an empty string and it can have sad side effects.
Implemented the configure_server(); should work although I did not test it...
Fixed the initialization of syslog with the name of the server program instead of the hard coded "snapserver".
SNAP-111: Applied a fix so we have a logger initialized early on (otherwise many errors may not make it there.)
SNAP-177: Added a logger_stub class so when no logging is going to happen we can avoid most of the code from doing anything.