EX-46: SNAP-67: Implemented the remaining code to support bouncing emails with Postfix.
Fixed the default levels in log.properties. It will be INFO by default.
snapbounce now saves the bounced emails in table "bounced_raw" because we do not yet have the website it is linked to. The sendmail plugin takes care of moving the matches in "bounced".
Added a parameter to the check_user_security() signal so we can choose to avoid blacklisted emails.
Check the email before using it in more cases. If the email is invalid we avoid sending anything to it.
Moved the test to know whether it is in our blacklists to the signal check_user_security() so the test happens whenever someone checks an email address.
Updated the session information for the email so we can determine the user email.
Updated the Message-ID to include the website key on top of the session identifier.
Fixed a potential lockup situation as I was calling the check_user_security() function within a block that locks the database.
Fixed the permissions for the hashtag settings.
Fixed the process_register_form() to avoid creating an account if the email validation fails. Instead we return an error to the user.
Changed the shorturl to shortlink as it looks like Google choose to use shortlink instead.
Added a startup plugin for Flash support.
Various clean ups.