EX-45: Finished up the implementation of the "cleanuplinks" function.
Fixed the direction of the 4 months test, it needed to be greater than instead of less than.
SNAP-172: I added support for an '*index*' row so we can access just the pages we need in one specific website since these columns are in a well defined order.
Fixed a bug with the dbutil with content::status_changed in the processing table.
Added MOVED as a status that the create_content() can restore to NORMAL.
Cleaned up on the backend of the links plugin.