SNAP-168: Added snapcommunicator to the snapserver package.
Created a snaputils with the tools such as cxpath, snapexpr, and snapimg, which are not really part of the snapserver.
Updated the apache2.conf with an example of port 443 setup.
Fixed the CGI directory in apache2.conf, we use the default and do not have to specify anything.
Also added an example that redirects users from HTTP to HTTPS.