SNAP-93: Implemented a dependencies() function in plugins so we can sort them without having to think too much about it. If you include A and B, then you depend on A and B. Easy.
The load works pretty much the same way, only it does register a bootstrap() signal, instead it is a virtual function now.
After the load, the first step is to sort the plugins by adding them to a vector and using their dependencies().
Then we execute the bootstrap() functions in that order. This means all the signals are in the correct order now, no need to rely on users entering the list of plugins in the correct order.
Changed the plugin name policy: it cannot start with a digit.
Updated all the plugins to the new scheme, especially, changed the on_bootstrap() to just bootstrap() and added the dependencies() function.
Many clean ups in the plugin implementation.