SNAP-168: Finished implementing the new snapinit tool.
Moved the SIGINT and SIGTERM to snap_signal so we can cleanly end the process in those cases.
Introduced the snapinit.xml file to describe each service.
Changed the --binary_path in --binary-path (underscore became dash.)
Fixed the snapcommunicator and snapinit configuration files.
Removed comment about having plugins defined in order in the default_plugins variable. Also about the 'no spaces'.
Correctly clear the g_log_output_filename when setting up a configuration file for the logger.
Correctly check stderr as a tty if we are to write there (instead of stdout).
Fixed a bug in the logger that would crash in is_enabled_for() if unconfigured. In that case we cannot access the g_logger and g_secure_logger variables.
Changed the signal handlers map in a sigset_t since we do not need a pointer back to the signal handlers (and smart pointers do not work in constructors).
Fixed the snap communicator connection comparison because we apply it to a list of shared pointers and not directly on snap_connection objects.
Changed the timeout delay so the minimum delay is 10. To get less, use a date.
Reimplemented the snap_signal class using the signalfd() function which allows us to wait on signals with poll().
Replaced the activate() and is_active() with a process() function which reads the signal info and calls the process_signal() as required.
Added a "Thread Done" signal using a pipe. This is used to immediately know (wake up a poll() call) that a thread is quitting.
Fixed the snap_tcp_client_buffer_connection which had to be marked as non_blocking() and wasn't.
Fixed the sort in the snap_communicator::run() function to a stable_sort().
Correctly clear the f_fds_position parameter to -1 on each loop. Only access the fds if that position is >= 0.
Fixed the snap_communicator loop so a loop with just a timeout works as expected.
Properly skip disabled connections when processing events.
Now snap_communicator proplery resets the timeout date once it ticked.
Moved the get_addr_port() function to tcp_client_server.
Added an msc description of the snapinit process since it uses a very large number of messages.
Added a left trim against the user agent so if it is just one or more spaces we also consider it empty.
Fixed the canonicalization of the neighbor IP:port specifications so it works with IPv6.
Fixed the parsing of messages in snapcommunicator as UDP really only supports STOP and SHUTDOWN at this time.
Added a thread_done_impl signal so when creating a remote socket we right away know once it is connected.
Got the snapcommunicator to work with snapinit.
Fixed the handling of connection types assuming that it is a multi-type field all the way.
Fixed the max_connections test from isEmpty() to not(isEmpty()).
Assume that we can still receive messages once we received a STOP since we are working in a multi-computer environment and can read messages from any socket.
Force the "types" parameter in ACCEPT and CONNECT, also check the "version" in CONNECT.
Fixed the register that was bogus since if service or version were not specified the execution would go on anyway.
Fixed the registration as the connection type was not set to "client" as expected.
Fixed the unregistration as the "service" field was expected to be mandatory but was not really.
Fixed the unregistration by setting the connection type to "" as expected.
Now a REGISTER is acknowledge with a READY. It is also were we send the HELP command.
Implemented the forwarding of messages to local services, including caching.
Implemented a SERVICES messages so snapinit, which has the authoritative list of services, can send that info to snapcommunicator.
Now reply to sending with "UNKNOWN" when a message we receive is not managed.
Fixed the up-since and down-since names as the dash is forbidden in message parameter names.
Support an empty string in add_neighbors().
Allow for a cache file in add_neighbors() so on following runs we can read the list of neighbors directly from the cache instead of just the ,conf file..
Actually send the STOP, DISCONNECT, SHUTDOWN message to the connection.
Fixed a crashing bug with snapcommunicator because the communicator was created on the stack and later we were trying to get a shared pointer.
Added snapinit and snapcommunicator to the snapsignal tool, even though it will change drastically once we have the proper implementation.
Renamed the snap communicator client test to be equivalent to what it is compiled as.
Various clean ups and documentation updates.