SNAP-168: Fixed many other bugs.
Added support for the LOG and UNKNOWN messages in snapinit.
Now we verify the name of commands, services, and parameters of messages. We also prevent a digit as the first character.
Fixed the retrieval of the address and port of the communicator when we want to send a STOP from the stop() function. It properly reads and uses the snapcommunicator.conf now.
Fixed the default port when using the snapcommunication date to 4041 instead of 4039.
Added the GOSSIP command so we can understand it. It is not yet implemented.
Implemented the LOG and QUITTING messages in the snapcommunicator.
Fixed the SERVICES message that was being refused when valid.
Added the UNKNOWN command to the snapcommunicator so we avoid some ping pong party.
Added support for broadcasting a message to all local services using "*" as the service name.
Fixed a bug in the shutdown() function that would make use of a reference of the list of connections which gets modified since we may remove some of them.
Properly remove the f_thread_done connection when we receive the event.
Fixed a bug in the snapcommunicator where the enable() function would enable entries that were not enabled before poll() cause potential problems with timeout callbacks that are not controlled by a flag that poll returns.
Fixed the properties by replacing RollingFileAppender by a standard FileAppender in an attempt to fix the multi-process bugs we are experiencing now.
Fixed te bug in the set_log_output_level() which would crash if called too early.
Added a reduce_log_output_level() so the --debug works as expected against log entries that use TRACE as their level.
Fixed the log levels in the logger_stub (it probably does not matter, but for cleanliness.)