SNAP-168: Created a "permanent connection" which is a TCP message connection that auto-reconnects itself when it gets lost.
Implemented and started to use the SIGCHLD connection in snapserver. This allows us to remove the zombies as soon as possible.
Implemented the TCP connection (which will replace the UDP connection) in snapserver. Not tested yet (since it is not used yet).
Fixed the .properties files some more as they where not all including the necessary Append flag.
Added a variable named snapcommunicator_listen to the snapserver.conf file.
Made logging::is_enabled_for() a global function instead of a static one.
Fixed a potential bug in snap_child::check_status() which could kill a zombie and not set the PID back to zero.
Allows the initialization of a snap_communicator::snap_timer with a timeout of zero. In that case it sets up the timer as a one time immediate tick.
Offer to retrieve the child PID on a SIGCHLD signal.
Better handle error cases on TCP buffer objects.
Correctly send an UNKNOWN message from snapinit when it receives a message it does not know how to handle.
Applied some corrections to our thread handling in snapcommunicator; thread used to stay connected with other snapcommunicators. This will be switch to our new permanent connection object in snapcommunicator, though.
Various clean ups.