SNAP-282: Added a missing attribute to the editor widget attachment: force-path.
This was in link with fading out the form for the editor, although these changes do not really include any such changes.
Fixed the CMakeLists.txt for favicon to properly check the editor and layout files (xml/xsl.)
Removed the settings-form.xml file since we do not use anymore.
Fixed the favicon tabindex table.
Removed the #include of the form and the session_id_t since it will not be used with the editor.
Fixed the feed CMakeLists.txt because it was not testing the feed settings files.
Added an atom-xslt20.dtd file to verify the atom.xsl file, which was a good idea as I found problems.
Fixed several errors in atom.xsl, such as the user of xsl:copy instead of xsl:value-of.
Fixed the default focus of the feed settings. It was referencing an info field still.
Added the editor as a dependency of feed and OAuth2 plugins.
Renamed the feed-xslt20.dtd as rss-xslt20.dtd since it is specific to the RSS format.
Moved the email of the OAuth2 settings (the user to access the global API functionality) to the secret table instead of the revision table since we want emails to remain secure.
Removed the toolbar from the email field of OAuth2.
Fixed the OAuth2 error replies which have to return a JSON or XML document with specific fields and not a traditional HTML error.