SNAP-168: Finally implemented the snap_backend with snap_communicator. We are nearly complete.
I won't put the details, snap_backend was very heavily modified to implement the new scheme (i.e. everything is now event driven.)
This especially included completely getting rid of the thread used to listen on the UDP port for a STOP or PING signal.
Renamed new_snapinit.cpp as snapinit.cpp as the old version is not compatible anymore.
Changed the way we specify actions on the command line by adding a --cron-action for actions that stick (except for the "list" action.)
Switch to make use of snap_communicator only in the server, instead of that temporary timer and still a "UDP thread."
Updated my ./start script to understand -r to delete the lock if still present and use the correct --binary-path command line option name
Enhanced the action registration using an actual class so we can verify that two actions with the same name are never used also we now add the name of the plugin as a namespace for each action.
Reduced the default wait on snapcommunicator to just one second instead of 10.
Updated dbutils.cpp to properly show content of the backend table.
Changed the call to the backend_process() signal, we can actually start it by calling an action callback. It is now in the content. That way we avoid special cases in the snap_backend.
Changed the implementation of the PING signal via UDP using the snap_communicator, we specify a service and a URI now.
Removed the get_signal_name() functions since all UDP signals are sent to the snap_communicator which in turn knows where to forward the events.
Enhanced the snap_communicator with the connection_added() and connection_removed() callbacks so we can do further initialization or destruction.
Added a snap communicator pipe connection so as to connect two processes (parent/child) that way they can communicate fully.
Moved all the code for backends that had a loop to the snap_backend implementation so plugins do not have to deal with that at all.
Fixed a few function of connection implementations so they properly call their base implementation.
Fixed a couple of functions which would return true even when the socket was closed.
Fixed a few cases where an error occurs by calling the process_error() function.
Fixed the messager as using dup() is not correct on a BIO file descriptor: we need to use the BIO_read() and BIO_write() calls on an SSL connection, so we have to keep that object around.
Fixed the thread_done() function that would create the messager client connection but not add it to the snap_communicator object.
Fixed a problem within stopping the permanent message connection, we have to let it know that we expect a HUP or it will continue to try to reconnect forever.
Fixed the fact that the permanent message implementation would reconnect to the snapcommunicator once a minute. We had to stop the timer once connected.
Fixed a bug with BIO_free_all() which actually calls shutdown() on the socket. We close the socket under its feet so that way we avoid the shutdown() (yuck!).
Fixed a bug with read and write: we must read from the BIO_read() and BIO_write(), we cannot just directly read/write with read() and write().
Fixed a bug in snapcommunicator that would not disconnect from children when receiving UNREGISTER. Yet that is required for the child to know it has to quit.
Enhanced the read()/write() functions that check more parameters for validity.
Fixed the get_addr() as the buffer would overflow in some cases.
Enhanced the get_addr_port() so it handles the case of the empty string and view both addr and port as defined with defaults on entry. It also understands service name to port now.
Fixed the snap_config [] operator so no new configuration parameter is created if used with the name of an undefined entry. Instead we just get an empty string. This means contains() will continue to return false on that entry.
Changed the snapsignal.cpp as a way to send messages to snapcommunicator as the new method is expected to work.
Removed the "quarterly" from sitemap.xml files since it is not actually valid.
Changed the order in which the tags are added, the priority was at the wrong place according to the .xsd validation file.
Added a client/server test to prove that shutdown() breaks the socket in parent or child.
Fixed some initialization in snap.cgi, although it is properly done in our existing constructor, for the future, it will be much safer.
Various clean ups.