SNAP-181: Adding a first version of snapfirewall to block IP addresses by just calling one function.
Added an entry in the snapinit.xml to start the snapfirewall.
Added a .conf and .properties and the necessary logrotate entries.
Added support for the "SAFE" message in snapinit so we can capture that signal from snapfirewall and start the snapserver and backends afterward.
Made sure dbutils knows the column names are int64 dates.
Added proper support for --debug in snapinit.
Added code to the snap_cassandra so we can create tables from that class. Actually, all cassandra accesses should be moved in that one class...
Fixed a bug in the snap_communicator thread done implementation: the process_read() function of the base class has to be called or the data stays in the pipe and thus the pipe cannot be used more than once.
Fixed a bug with the messager thread: the thread done connection was never added to the communicator, meaning that it would never trigger anything.
Now that we properly add the thread done signal in the messager, we also properly remove it once the trigger was received.
Added a static function to the server that will allow most everyone to send a BLOCK message to the snapfirewall in order to block an IP address.
Implemented the block_ip() in our snap.cgi which already detects some edge cases of invalid and thus hackish connections.
Updated the snapdb.cpp to drop tables "backend" and "firewall".