Updated the QR Code to support an option on whether to use the Short URL or not.
Added a flag to allow people to use a query string (?qrcode=true) in the QR Code so one can track such hits.
Put a comment in the PNG we generate so one can see the URL without a decoder.
Avoid attempting to create a QR code for the "qrcode-not-available.png" image. Really no need.
Added the editor as an explicit dependency.
Various clean ups.