SNAP-109: Got the snapwatchdogserver to work with the latest of everything.
Fixed the snapwatchdog.properties with "snapwatchdog.log" instead of "snapserver.log".
Actually implemented the sigchld_connection in there. It was not created/used yet.
Fixed the process_tick() the test of f_processes was inverted.
Got the STOP signal to work as expected.
Fixed a bug in the get_server_name(), one was misspelled.
Made enumerations classes.
Various other clean ups.