SNAP-289: Added tests for return values that were ignored in many places.
Fixed the NOTUSED to make use of a const reference instead of by copy. Since it gets compiled into nothing, it certainly makes no difference, but Coverity mentioned it to us.
Changed some controlled_vars::r<type> (random) to use an initialization to make Coverity happy.
Verify a certain number of dynamic_cast<>() properly.
Fixed a couple of function where the bio_client would use the socket even if -1.
Fixed a few bugs in the csspp library as shown by Coverity.
Fixed a bug in the images plugin where the "copy" test had to have an else before the if().
SNAP-168: Started tweaks to get multi-computer cluster function as expected.
Changed implementation so remote connections make use of a permanent connection (we had such, only it had some bugs and we should not duplicate such code.)
Added get_addr()/get_address() to the permanent connection.
Fixed the get_addr()/get_address() because we were not using the set_addr[ess]() anymore; now we get the data from the socket.
Implemented the local_listen for local services, opposed to the main listen for other snapcommunicator to connect to.
Added support for "my_address" so I know which IP I'm on and can use that information to connect to remote snapcommunicators (i.e. if IP is smaller, I connect, if larger, they connect, if equal, bad setup...)
Properly mark the server as done (f_messager) when we receive the STOP or QUITTING message and send the UNREGISTER as expected.
Check the socket for -1 when we accept() new connections in snap server.
Clarified some objects by choosing a "better" (hopefully) name for them.
Fixed the snapserver so it quits cleanly (unregistering and waiting for the HUP as expected.)
Added a new addr class to handle IP addresses (finally!) I would have to convert all address handling to using that class now...
Various clean ups.