SNAP-289: Applied a few fixes for issues found with Coverity.
Fixed a potential crash bug in the lexer::ungetc(), the limit was checked against the buffer size in bytes instead of number of entries.
Fixed a test in a while() that was missing parenthesis.
Compiler created in lib/compiler.cpp was using 'this' instead of 'true'.
Marked with a comment all cases in switches which are expected to fall through.