EX-13: Timeout the user session by implementing an auto-logout on the client side.
Added a Users Settings page to edit the session parameters that the administrator can change.
Added documentation about the different type of session times so it is cleared in my head too.
Added support so an administrator can change the login administrative session to a soft session (i.e. extended each time you hit the server.)
Added a flag named f_administrative_logged_in which allows us to give additional rights to people (i.e. change their password, go to /admin...)
Finally added support for inline JavaScript. In this case it is to add the timeout dates of the users' session.
Fixed a bug in the cookie implementation, a computation was done in 32 bit when it needed to be 64 bit. The expiration date is now working as expected.
In the set_expire[_in](), we now make sure to clamp the time to at most 1 year.
Fixed the cookie date, GMT had to be written between single quote, otherwise the M is transformed with the month.
Fixed the antivirus that used the word antivirus unquoted in the QDateTime::toString() function.
The session_info now accepts the remote IP address as a parameter.
Fixed a bug in the sessions where in one case I was not caculating the correct TTL for a value being added to the sessions table.
Fixed the session TTL calculation and moved it in a function so we can use it in various places.
Save when the session is created so that way we can determine its age.
Added a permission for the retruning visitor, opposed to a visitor (i.e. whether you have an old session or not.)
Various clean ups.