EX-6: SNAP-161: Editor loads the original body and title to properly edit the data with tokens and no other filtering (i.e. adding icon for external links, auto-table of contents, etc.)
Fixed the bare theme body-parser.xsl file as the name was missing in one of the div.
Added the breadcrumbs to the bare theme.
Made the title of a page editable in the bare theme.
Made the column type available publicly in the dbutils class.
Fixed the CSS of the line-edit in core. The "white-space: nowrap" + "overflow: hidden" were not defined.
Moved the definition of the serverAccess_ field in the base class of the editor widget type tree.

SNAP-291: Implemented a first version to Install and Remove plugins.
Fixed the list_all() function so only the files that end with .so are returned (and not directories!) The function also reads all sub-directories (one level.)
Plugins now can define icon(), plugin_categorization_tags(), help_uri(), and settings_path().
Added a plugin_info() class so we can actually retrieve all the necessary data when listing plugins.
Added a is_core_plugin() function to know whether a plugin is a core plugin or not (as defined by the minimum default set of plugins in snap_child.)
Fixed some of the settings pages so we can at least access them.

EX-13: Added a "hits" query string to determine whether a user hit has to be taken in account in regard to extending the user session or not.

Added the necessary functionality to support the "hit" variable in the sserver-access plugin.
Applied a fix so the user session does not get extended once it timed out. Otherwise the user is automatically re-logged in (ha! ha! what a joke that was!)

Updated the snapserver.conf to specify that the plugins_path variable actually supports multiple paths when separated by colons.
Added a DOM helper to add an integer to a a node (instead of having to do QString("%1").arg(int) each time.)
Fixed the editor text edit boxes: the arrows still work even when the field is disabled or marked read-only.
Added support for executing the token filtering over XSLT files by adding a 'filter="token"' attribute to their output tag.
Fixed the code adding the site_name, now it forces a default if still undefined, it looks "better" that way.
Check the request URI case insensitively in snap.cgi and added a test against the name "phpmyadmin" which will bounce right away and block the user.
Strengthen the test of the user agent as it is often set to "-". Also reject "ZmEu" requests.

Various clean ups.