SNAP-259: Created password form for the "users" and other policies.
Added support for multiple password policies (i.e. "users", "protected page", etc.)
Created a new create_password() function that takes the policy in account to generate the password.
Fixed the register_user() call in sendmail by testing the returned value and acting appropriately, especially if the user is blocked and the function fails completely.
Changed the check_user_security() signal to use one object instead of 5 variables and added functionality so we can distinguish errors such as an invalid password versus an unwanted email.
Added a case when the password is considered too weak in the user_ui registration handling.
Added fields to dbutils.cpp.

SNAP-269: Added a password_confirm editor widget to ask users to enter a password twice (JS not complete yet).
Added a background-confirm field for the new password_confirm widget.
Removed the required "foo" attribute for the attachment, background-value, help, and tooltip in the XSLT 2.0 DTD.

SNAP-161: Applied a fix to get the correct size/format for the "Edit" button.

SNAP-168: Fixed a bug in snap_backend.cpp where I was still displaying f_cron_action after I changed it to a bool.

SNAP-291: Made "locale" part of the core plugins since "users" accesses it.
Fixed a problem with the list of plugins when starting with an empty slate.

Various clean ups.