SNAP-269: Added form to test/add/remove blacklisted passwords.
Fixed the source path to strip in the doxy file.
Removed a spurious comma from editor.css.
Made the dropped file area a minimum of 200x200 pixels by default.
Fixed a bug in the editor.js file. If a an ImageBox widget does not have a background value, it does not try to change the position anymore.
Various cleanups.