SNAP-293: Password timeout implemented. When a user logs in, if his password timed out, he is forced to enter a new password.
Also implemented the password history so a user cannot just reuse the same password over and over again.
Added support for a minimum variation instead of just a set of minimum of specific types.
Added an OAuth2 password policy.
Added an strstr_with_levenshtein_distance() function to search a string in another with the levenshtein distance to compare both strings.
Changed the OAuth2 login to use the OAuth2 policy.
Fixed a problem with the on_check_user_security() which may be called with an empty email.
Added the name of the password policy to the user_logged_info_t object.
Moved the save_password() signal to the users.cpp instead of three duplications in users.cpp and users_ui.cpp.
Moved the user_logged_info_t implementation to its own file.
Actually added the calls to the check_user_security() to the users_ui.cpp when the user is trying to change his password.